@@ -1,6 +1,6 @@
 {
   "name": "jon-template",
-  "version": "0.2.0",
+  "version": "0.5.0",
   "private": false,
   "//dependencies": {
     "@priolo/jon#local": "file:C://Users//iorio//Documents//Jon",
@@ -14,7 +14,6 @@
     "@emotion/styled": "^11.8.1",
     "@mui/icons-material": "^5.6.1",
     "@mui/material": "^5.6.1",
-    "@mui/styles": "^5.6.1",
     "@priolo/jon": "file:C://Users//iorio//Documents//Jon",
     "@priolo/jon-utils": "file:C://Users//iorio//Documents//jon-utils",
     "i18next": "^21.6.11",
@@ -22,11 +21,11 @@
     "i18next-http-backend": "^1.3.2",
     "js-cookie": "^2.2.1",
     "msw": "^0.27.1",
-    "react": "17",
-    "react-dom": "17",
-    "react-scripts": "4",
+    "react": "18",
+    "react-dom": "18",
+    "react-scripts": "5",
     "react-i18next": "^11.16.6",
-    "react-router-dom": "^5.2.0"
+    "react-router-dom": "6"
   },
   "devDependencies": {
     "cross-env": "^7.0.3",
