@@ -1,21 +1,21 @@
 import React, { useState } from 'react';
 
-import makeStyles from '@mui/styles/makeStyles';
+// mui
 import { AccountCircle as AccountIcon, ExitToApp as LogoutIcon, Face as ProfileIcon } from '@mui/icons-material';
 import { IconButton, Menu, MenuItem, Typography, Box, Divider, ListItemIcon, Switch } from '@mui/material';
-
+// 
 import { useTranslation } from 'react-i18next';
-import { useHistory } from "react-router-dom";
-
+import { useNavigate } from "react-router-dom";
+// stores
 import layoutStore from "stores/layout";
 import authStore from "stores/auth";
 import { useStore17 } from "@priolo/jon";
 
+
 function Avatar() {
 
 	// HOOKs
-	const classes = useStyles()
-	const history = useHistory()
+	const navigate = useNavigate()
 	const [anchorEl, setAnchorEl] = useState(null)
 	const { t } = useTranslation()
 
@@ -23,7 +23,6 @@ function Avatar() {
 	const { logout } = authStore
 	useStore17(layoutStore)
 	const { toggleTheme, isDarkTheme } = layoutStore
-
 
 	// HANDLEs
 	const handleClose = () => setAnchorEl(null)
@@ -34,9 +33,8 @@ function Avatar() {
 	}
 	const handleClickProfile = e => {
 		handleClose()
-		history.push("/profile")
+		navigate("/profile")
 	}
-
 
 	// RENDER
 	return <>
@@ -50,7 +48,7 @@ function Avatar() {
 			open={Boolean(anchorEl)}
 			onClose={handleClose}
 		>
-			<Box className={classes.box}>
+			<Box sx={cssBox}>
 				<Typography variant="subtitle2">{auth.user.email}</Typography>
 				<Typography variant="caption">({t(`app.roles.${auth.user.role}`)})</Typography>
 			</Box>
@@ -80,14 +78,12 @@ function Avatar() {
 				{t("app.avatar.logout")}
 			</MenuItem>
 		</Menu>
-	</>;
+	</>
 }
 
 export default Avatar
 
-const useStyles = makeStyles((theme) => ({
-	box: {
-		padding: theme.spacing(1, 2, 1, 2),
-		display: "flex", flexDirection: "column"
-	}
-}));
+const cssBox = theme => ({
+	padding: theme.spacing(1, 2, 1, 2),
+	display: "flex", flexDirection: "column"
+})
