@@ -1,16 +1,12 @@
 import React from "react"
-import ReactDOM from "react-dom"
+import { createRoot } from "react-dom/client";
 import './index.css';
 import Main from "./components/layouts/Main"
-
 
 import './plugins/msw';
 import './plugins/i18n';
 
 
 // APPLICATION
-const rootElement = document.getElementById("root")
-ReactDOM.render(
-	<Main />,
-	rootElement
-)
+const root = createRoot(document.getElementById('root'))
+root.render(<React.StrictMode><Main /></React.StrictMode>)
