@@ -1,4 +1,3 @@
-import makeStyles from '@mui/styles/makeStyles';
 import { Search as SearchIcon, Clear as ClearIcon } from '@mui/icons-material';
 import { InputBase, alpha, IconButton, Box } from '@mui/material';
 
@@ -9,8 +8,6 @@ function SearchBox({
 }) {
 
 	// HOOKs
-	const classes = useStyles()
-	
 
 	// HANDLEs
 	const handleChange = e => onChange(e?.target.value ?? "")
@@ -18,57 +15,53 @@ function SearchBox({
 
 	// RENDER
 	const haveAdorner = value != null && value.length > 0
-	
-	return (
-        <div className={classes.searchBox}>
 
-			<SearchIcon className={classes.searchIcon} />
+	return (
+		<Box sx={cssSearchBox}>
+
+			<SearchIcon sx={cssSearchIcon} />
 
 			<Box ml={2} />
 
 			<InputBase
 				placeholder="Search…"
-				className={classes.input}
+				sx={cssInput}
 				value={value}
 				onChange={handleChange}
 				endAdornment={haveAdorner && (
-					<IconButton className={classes.clearIcon} onClick={e => handleChange()} size="large"><ClearIcon /></IconButton>
+					<IconButton sx={cssClearIcon} onClick={e => handleChange()} size="large"><ClearIcon /></IconButton>
 				)}
 			/>
 
-		</div>
-    );
+		</Box>
+	);
 }
 
 export default SearchBox
 
-const useStyles = makeStyles((theme) => ({
-
-	searchBox: {
-		display: "flex",
-		alignItems: "center",
-		width: "100%",
-		minHeight: "48px",
-		padding: "0px 10px",
-		borderRadius: theme.shape.borderRadius,
-		backgroundColor: alpha(theme.palette.common.white, 0.15),
-		'&:hover': {
-			backgroundColor: alpha(theme.palette.common.white, 0.25),
-		},
+const cssSearchBox = theme => ({
+	display: "flex",
+	alignItems: "center",
+	width: "100%",
+	minHeight: "48px",
+	padding: "0px 10px",
+	borderRadius: theme.shape.borderRadius,
+	backgroundColor: alpha(theme.palette.common.white, 0.15),
+	'&:hover': {
+		backgroundColor: alpha(theme.palette.common.white, 0.25),
 	},
+})
 
-	searchIcon: {
-		flex: "0 1 auto",
-	},
+const cssSearchIcon = theme => ({
+	flex: "0 1 auto",
+})
 
-	clearIcon: {
-		color: "inherit",
-		padding: "0 5px"
-	},
+const cssClearIcon = theme => ({
+	color: "inherit",
+	padding: "0 5px"
+})
 
-	input: {
-		flex: "1 1 auto",
-		color: 'inherit',
-	},
-
-}));
+const cssInput = theme => ({
+	flex: "1 1 auto",
+	color: 'inherit',
+})
