@@ -1,7 +1,7 @@
 import React from 'react'
 import { Grid } from "@mui/material";
-import makeStyles from '@mui/styles/makeStyles';
 import CentralSpace from '../layouts/CentralSpace'
+import { Box } from '@mui/system';
 
 
 export default function Form({
@@ -10,58 +10,76 @@ export default function Form({
 	renderMenu,
 }) {
 
-	// HOOKs
-	const classes = useStyles()
-
-
 	// RENDER
 	return (
-		<div className={classes.frame}>
+		<Box sx={cssFrame}>
 
 			<CentralSpace
-				className={classes.content}
+				sx={cssContent}
 				renderRight={renderMenu && (
-					<div className={classes.menu}>
+					<Box sx={cssMenu}>
 						{renderMenu}
-					</div>
+					</Box>
 				)}
 			>
 				{children}
 			</CentralSpace>
 
 			{renderFooter && (
-				<CentralSpace className={classes.footer}>
+				<CentralSpace sx={cssFooter}>
 					<Grid container>{renderFooter}</Grid>
 				</CentralSpace>
 			)}
 
-		</div>
+		</Box>
 	)
 }
 
+const cssFrame = theme => ({
+	flex: "1 1 auto", display: "flex", flexDirection: "column",
+})
+const cssContent = theme => ({
+	flex: "1 1 auto",
+})
+const cssMenu = theme => ({
+	display: "flex", flexDirection: "column",
+	position: "sticky", top: "64px",
+	padding: "49px 16px 16px 0px", margin: "0 0 0 40px",
 
-const useStyles = makeStyles(theme => ({
-	frame: {
-		flex: "1 1 auto", display: "flex", flexDirection: "column",
-	},
+})
+const cssFooter = theme => ({
+	borderTop: "#e0e0e0 1px solid",
+	padding: "18px 0",
+	display: "flex", flexDirection: "row", alignItems: "center",
+	position: "sticky", zIndex: "1", bottom: 0,//`${theme.app.footerbar.height}px`,
+	height: "70px",
+	backgroundColor: theme.palette.background.default,
+})
 
-	content: {
-		flex: "1 1 auto",
-	},
-	// body: {
-	// 	marginBottom: "50px",
-	// },
-	menu: {
-		display: "flex", flexDirection: "column",
-		position: "sticky", top: "64px",
-		padding: "49px 16px 16px 0px", margin: "0 0 0 40px",
-	},
-	footer: {
-		borderTop: "#e0e0e0 1px solid",
-		padding: "18px 0",
-		display: "flex", flexDirection: "row", alignItems: "center",
-		position: "sticky", zIndex: "1", bottom: 0,//`${theme.app.footerbar.height}px`,
-		height: "70px",
-		backgroundColor: theme.palette.background.default,
-	},
-}));
+
+
+// const useStyles = makeStyles(theme => ({
+// 	frame: {
+// 		flex: "1 1 auto", display: "flex", flexDirection: "column",
+// 	},
+
+// 	content: {
+// 		flex: "1 1 auto",
+// 	},
+// 	// body: {
+// 	// 	marginBottom: "50px",
+// 	// },
+// 	menu: {
+// 		display: "flex", flexDirection: "column",
+// 		position: "sticky", top: "64px",
+// 		padding: "49px 16px 16px 0px", margin: "0 0 0 40px",
+// 	},
+// 	footer: {
+// 		borderTop: "#e0e0e0 1px solid",
+// 		padding: "18px 0",
+// 		display: "flex", flexDirection: "row", alignItems: "center",
+// 		position: "sticky", zIndex: "1", bottom: 0,//`${theme.app.footerbar.height}px`,
+// 		height: "70px",
+// 		backgroundColor: theme.palette.background.default,
+// 	},
+// }));
