@@ -1,8 +1,6 @@
 import React from 'react';
 import { CardHeader, Card, CardContent, CardActions, CircularProgress, Divider } from '@mui/material';
-import makeStyles from '@mui/styles/makeStyles';
 import { Skeleton } from '@mui/material';
-
 
 export default function FormParagraph({
 	id,
@@ -13,27 +11,22 @@ export default function FormParagraph({
 	bodyBusy,
 	haveExpand,
 }) {
-
-	// HOOKs
-	const classes = useStyles();
-
-
 	// RENDER
 	return (
-		<Card className={classes.root} id={id}>
+		<Card sx={cssRoot} id={id}>
 			<CardHeader title={title} />
 
-			<CardContent className={haveExpand ? classes.content : null}>
+			<CardContent sx={haveExpand ? cssContent : null}>
 				{bodyBusy
 					? <Skeleton animation="wave" variant="rectangular" height={200} />
 					: children
 				}
 			</CardContent>
 			{renderFooter && (<>
-				<Divider className={classes.divider} />
-				<CardActions className={classes.actions}>
+				<Divider sx={cssDivider} />
+				<CardActions sx={cssActions}>
 					{footerBusy
-						? <CircularProgress size={36} className={classes.progress} />
+						? <CircularProgress size={36} sx={cssProgress} />
 						: renderFooter
 					}
 				</CardActions>
@@ -43,24 +36,22 @@ export default function FormParagraph({
 }
 
 
-const useStyles = makeStyles((theme) => ({
-	root: {
-		padding: "9px 16px 16px 16px",
-		marginTop: theme.spacing(3),
-		marginBottom: theme.spacing(3),
-		width: "100%",
-	},
-	content: {
-		paddingBottom: "0px!important"
-	},
-	actions: {
-		//marginTop: "10px",
-		padding: "16px",
-	},
-	progress: {
-		marginLeft: "8px",
-	},
-	divider: {
-		margin: "9px -17px;",
-	}
-}));
+const cssRoot = theme => ({
+	padding: "9px 16px 16px 16px",
+	marginTop: theme.spacing(3),
+	marginBottom: theme.spacing(3),
+	width: "100%",
+})
+const cssContent = theme => ({
+	paddingBottom: "0px!important"
+})
+const cssActions = theme => ({
+	//marginTop: "10px",
+	padding: "16px",
+})
+const cssProgress = theme => ({
+	marginLeft: "8px",
+})
+const cssDivider = theme => ({
+	margin: "9px -17px;",
+})
