@@ -1,78 +1,76 @@
-import makeStyles from '@mui/styles/makeStyles';
-
 import layoutStore from "stores/layout";
 import { useStore17 } from "@priolo/jon";
+import { Box } from '@mui/system';
 
 
 export default function CentralSpace({
 	renderLeft,
 	renderRight,
 	children,
-	className,
+	sx,
 	isCentered = false,
 }) {
 
 	// HOOKs
-	const classes = useStyles()
 	const layout = useStore17(layoutStore)
 
-
 	// RENDER
-	const cnContainer = `${classes.container} ${className ?? ""} ${isCentered ? "centered" : ""}`
-	const cnLateral = `${classes.lateral} ${layout.device}`
-	const cnCentral = `${classes.central} ${layout.device}`
-	
+	const sxContainer = theme => cssContainer(theme, sx, isCentered)
+	const sxLateral = theme => cssLateral(theme, layout.device)
+	const sxCentral = theme => cssCentral(theme, layout.device)
+
 	return (
-		<div className={cnContainer}>
+		<Box sx={sxContainer}>
 
-			<div className={cnLateral}>
+			<Box sx={sxLateral}>
 				{renderLeft}
-			</div>
+			</Box>
 
-			<div className={cnCentral}>
+			<Box sx={sxCentral}>
 				{children}
-			</div>
+			</Box>
 
-			<div className={cnLateral}>
+			<Box sx={sxLateral}>
 				{renderRight}
-			</div>
+			</Box>
 
-		</div>
+		</Box>
 	)
 }
 
-const useStyles = makeStyles(theme => ({
 
-	container: {
-		display: "flex",
-		width: "100%",
-		"&.centered": {
-			alignItems: "center"
-		}
-	},
 
-	// lateral space
-	lateral: {
-		display: "flex", flex: "1 1 0%",
-		"&.mobile": {
+const cssContainer = (theme, parentSx, isCentered) => ({
+	display: "flex",
+	width: "100%",
+	...parentSx?.(theme),
+	...(isCentered && { alignItems: "center"})
+})
+
+// lateral space
+const cssLateral = (theme, device) => ({
+	display: "flex", flex: "1 1 0%",
+	...({
+		"mobile": {
 			display: "flex", flex: "0 1 5px",
 		},
-		"&.pad": {
+		"pad": {
 			display: "flex", flex: "0 1 10%",
-		}
-	},
+		},
+	}[device])
+})
 
-	// central space
-	central: {
-		flex: "3 1 0%",
-		minWidth: "300px", 
-		maxWidth: "700px",
-		"&.mobile": {
+// central space
+const cssCentral = (theme, device) => ({
+	flex: "3 1 0%",
+	minWidth: "300px",
+	maxWidth: "700px",
+	...({
+		"mobile": {
 			maxWidth: "unset",
 		},
-		"&.pad": {
+		"pad": {
 			maxWidth: "unset",
-		}
-	},
-
-}))
+		},
+	}[device])
+})
