@@ -1,6 +1,4 @@
-/* eslint eqeqeq: "off", react-hooks/exhaustive-deps: "off"*/
-import makeStyles from '@mui/styles/makeStyles';
-import { Divider, Drawer, IconButton, List } from "@mui/material"
+import { Divider, Drawer, IconButton, List, Box } from "@mui/material"
 import { ChevronLeft as ChevronLeftIcon } from "@mui/icons-material"
 
 import ItemMenu from "./ItemMenu"
@@ -13,25 +11,21 @@ function MainDrawer() {
 	// HOOKs
 	const layout = useStore17(layoutStore)
 	const { toggleDrawerIsOpen, getDrawerList } = layoutStore
-	const classes = useStyles()
-
 
 	// RENDER
 	const variant = layout.device == "desktop" ? "persistent" : null
 
 	return (
-		<Drawer
+		<Drawer sx={cssDrawer}
 			variant={variant}
-			className={classes.drawer}
 			open={layout.drawerIsOpen}
-			classes={{ paper: classes.drawerPaper }}
 			ModalProps={{ onBackdropClick: toggleDrawerIsOpen }}
 		>
-			<div className={classes.drawerHeader}>
+			<Box sx={cssDrawerHeader}>
 				<IconButton onClick={toggleDrawerIsOpen} size="large">
 					<ChevronLeftIcon />
 				</IconButton>
-			</div>
+			</Box>
 			<Divider />
 			<List>
 				{getDrawerList().map(item => (
@@ -44,20 +38,16 @@ function MainDrawer() {
 
 export default MainDrawer
 
-const useStyles = makeStyles((theme) => ({
-	drawer: {
-		width: theme.app.drawer.width,
-		flexShrink: 0
-	},
-	drawerPaper: {
-		width: theme.app.drawer.width
-	},
-	drawerHeader: {
-		display: "flex",
-		alignItems: "center",
-		padding: theme.spacing(0, 1),
-		// necessary for content to be below app bar
-		...theme.mixins.toolbar,
-		justifyContent: "flex-end"
-	}
-}))
+const cssDrawer = theme => ({
+	width: theme.app.drawer.width,
+	flexShrink: 0,
+	"& .MuiDrawer-paper": {width: theme.app.drawer.width},
+})
+const cssDrawerHeader = theme => ({
+	display: "flex",
+	alignItems: "center",
+	padding: theme.spacing(0, 1),
+	// necessary for content to be below app bar
+	...theme.mixins.toolbar,
+	justifyContent: "flex-end"
+})
