@@ -1,8 +1,6 @@
-/* eslint eqeqeq: "off", react-hooks/exhaustive-deps: "off"*/
 import { useEffect } from "react";
 import { useTranslation } from "react-i18next";
 import { Drawer, Grid, IconButton, Typography } from "@mui/material";
-import makeStyles from '@mui/styles/makeStyles';
 import { Close as CloseIcon } from '@mui/icons-material'
 
 import DocFilters from "pages/doc/DocFilters";
@@ -16,13 +14,12 @@ import { useStore17 } from "@priolo/jon";
 function RightDrawer() {
 
 	// HOOKs
-	const classes = useStyles()
 	const { t } = useTranslation()
 
 	const layout = useStore17(layoutStore)
-	const {setDrawerRightIsOpen} = layoutStore
+	const { setDrawerRightIsOpen } = layoutStore
 	const route = useStore17(routeStore)
-	const {haveSearchExtra} = routeStore
+	const { haveSearchExtra } = routeStore
 
 	const haveExtra = haveSearchExtra()
 
@@ -31,7 +28,7 @@ function RightDrawer() {
 		else if (haveExtra) setDrawerRightIsOpen(true)
 	}, [route.currentPage, haveExtra])
 
-	
+
 	// HANDLEs	
 	const handleClickClose = () => setDrawerRightIsOpen(false)
 
@@ -42,10 +39,10 @@ function RightDrawer() {
 	}[route.currentPage] ?? null
 
 	return (
-        <Drawer anchor="right" variant="persistent"
+		<Drawer anchor="right" variant="persistent"
 			open={layout.drawerRightIsOpen}
 		>
-			<Grid container direction="column" className={classes.container}>
+			<Grid container direction="column" sx={cssContainer}>
 
 				<Grid item container alignItems="center">
 					<Grid xs item>
@@ -60,15 +57,12 @@ function RightDrawer() {
 
 			</Grid>
 		</Drawer>
-    );
+	);
 }
 
 export default RightDrawer
 
-const useStyles = makeStyles(theme => ({
-	container: {
-		margin: `${theme.app.header.height+10}px 15px 10px 15px`,
-		width: theme.app.drawer.width,
-	}
-	
-}));
+const cssContainer = theme => ({
+	margin: `${theme.app.header.height + 10} 15px 10px 15px`,
+	width: theme.app.drawer.width,
+})
